Add Formatter trait for validating & formatting text
This is a mechanism for handling conversions back and forth
between values and their textual representations. Its major
use case is managing user input in text fields when the input
is intended to represent some other value type.

This patch includes the trait, a simple implementation based
on `FromStr` (intended as a replacement for the `Parse` widget)
and a new ValueTextBox widget that uses some Formatter to manage
a TextBox.

The behaviour of the ValueTextBox differs from the normal TextBox
in one very important way; it does not attempt to update the data
until editing has 'completed'. Editing is completed when the user
presses <return> or attempts to navigate away from the textbox;
the user can also manually cancel editing by pressing <esc>, which
discards the edit.

These interactions (especially the hard-wired key handling) feels
a bit brittle, and I'd like to try and find a more systematic
approach.

This is still a draft; there are a number of rough edges, and
other things I'm not totally happy with, including:

- it uses String everywhere; I might prefer Arc<String> at least
- there is no good mechanism for actually reporting errors back
up the tree
- the error type itself is a bit too fancy, and should be simplified
- I'd like to add a general purpose NumberFormatter
- I'd like to look into incorporating this into the existing TextBox,
including the idea of 'on_completion' and 'on_cancel' handling.

In any case, I'm happy with the progress here; at the very least this
is a much more robust and flexible alternative to the current Parse
widget.

I have updated `examples/parse.rs` to show off some of what is
possible with this API; that example should probably be renamed.